preload: add purely static LD_PRELOAD injection
Open a purely memory-backed file and write the libremain.so code into
it. Currently this does technically cause you to have a few kilobyte
file open in the child process's context. Unfortunately, doing this
trick means that we can't close the file. However, this is (in
principle) a one-time cost.

Signed-off-by: Aleksa Sarai <asarai@suse.de>